server/requirements.txt
fastapi
SQLAlchemy
uvicorn
python-dotenv
PyMySQL
python-jose
google-auth
pandas
requests
openai